# Handle records and maps that are OrderedDict rather than dict