--- a/fastavro/_write.pyx
+++ b/fastavro/_write.pyx
@@ -369,7 +369,7 @@ cpdef write_map(bytearray fo, object datum, dict schema):
     If a block's count is negative, then the count is followed immediately by a
     long block size, indicating the number of bytes in the block. The actual
     count in this case is the absolute value of the count written."""
-    cdef dict d_datum
+    cdef dict d_datum = None
     try:
         d_datum = <dict?>(datum)
 
@@ -382,6 +382,13 @@ cpdef write_map(bytearray fo, object datum, dict schema):
                 write_data(fo, val, vtype)
         write_long(fo, 0)
     except TypeError:
+        # If d_datum has a value, then 'datum' *is* a dictionary, so the
+        # TypeError occurred for some other reason, probably while writing
+        # data. In this case, bail out so we don't possibly corrupt the output
+        # file.
+        if d_datum is not None:
+            raise
+
         # Slower, general-purpose code where datum is something besides a dict,
         # e.g. a collections.OrderedDict or collections.defaultdict.
         if len(datum) > 0:
@@ -543,7 +550,7 @@ cpdef write_record(bytearray fo, object datum, dict schema):
     their schema."""
     cdef list fields
     cdef dict field
-    cdef dict d_datum
+    cdef dict d_datum = None
     fields = schema['fields']
     try:
         d_datum = <dict?>(datum)
@@ -557,6 +564,13 @@ cpdef write_record(bytearray fo, object datum, dict schema):
             write_data(fo, d_datum.get(
                 name, field.get('default')), field['type'])
     except TypeError:
+        # If d_datum has a value, then 'datum' *is* a dictionary, so the
+        # TypeError occurred for some other reason, probably while writing
+        # data. In this case, bail out so we don't possibly corrupt the output
+        # file.
+        if d_datum is not None:
+            raise
+
         # Slower, general-purpose code where datum is something besides a dict,
         # e.g. a collections.OrderedDict or collections.defaultdict.
         for field in fields: