Address #139.

We add a runtime check to `write_record()` and `write_map()`. For `dict`, use faster code. For other types, use equivalent but slower, generic code. There appears to be no meaningful impact on performance.

```
Before:

Writing 10000 records to one file...
... 10 runs averaged 0.00915930271149 seconds
Reading 10000 records from one file...
... 10 runs averaged 0.0144769668579 seconds

Writing 10000 records to one file...
... 10 runs averaged 0.0550956249237 seconds
Reading 10000 records from one file...
... 10 runs averaged 0.0886501789093 seconds

After:

Writing 10000 records to one file...
... 10 runs averaged 0.00916681289673 seconds
Reading 10000 records from one file...
... 10 runs averaged 0.0137780666351 seconds

Writing 10000 records to one file...
... 10 runs averaged 0.0546228170395 seconds
Reading 10000 records from one file...
... 10 runs averaged 0.0791359186172 seconds
```