@barrywhart 2017-12-27
I'm pretty sure the same technique could be extended to address #76 as well. Performance for the basic `dict` case would not be affected because the additional checks for `list` or `tuple` would happen if and only after the initial check for `dict` had failed.